@@ -1,1 +1,1 @@
-default_app_config = "course.apps.CourseConfig"
+# (empty)
